mv deps comparasion into reqs
