+++ dephell/models/dependency.py
@@ -4,7 +4,6 @@ from typing import Tuple
 
 # external
 import attr
-from dephell_links import VCSLink
 from dephell_markers import Markers
 from packaging.utils import canonicalize_name
 
@@ -22,6 +21,7 @@ class Dependency:
     constraint = attr.ib(type=Constraint)
     repo = attr.ib(repr=False)
     link = attr.ib(default=None, repr=False)
+    marker = None
 
     # flags
     applied = attr.ib(type=bool, default=False, repr=False)
@@ -262,18 +262,3 @@ class Dependency:
 
     def __lt__(self, other: 'Dependency') -> bool:
         return self.name < other.name
-
-    @staticmethod
-    def _get_comparable_dict(dep) -> dict:
-        excluded = {'constraint', 'repo', 'link', 'marker', 'license', 'inherited_envs', 'locations'}
-        result = attr.asdict(dep, recurse=True, filter=lambda x, _: x.name not in excluded)
-        result['constraint'] = str(dep.constraint)
-        if dep.marker:
-            result['marker'] = str(dep.marker)
-        if isinstance(dep.link, VCSLink):
-            result['link'] = attr.asdict(dep.link, recurse=True)
-        result['license'] = str(dep.license)
-        return result
-
-    def __eq__(self, other) -> bool:
-        return self._get_comparable_dict(self) == self._get_comparable_dict(other)
